made class for projectile
