+++ Spacekitten/GameScene.swift
@@ -63,6 +63,7 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
     let player = SKSpriteNode(imageNamed: "red")
     let hud = HUD()
     
+    
     var enemiesDestroyed = 0
     var playerSize = 120
     
@@ -184,50 +185,21 @@ class GameScene: SKScene, SKPhysicsContactDelegate {
     
     override func touchesEnded(touches: Set<UITouch>, withEvent event: UIEvent?) {
         
-        // 1 - Choose one of the touches to work with
+        // Choose one of the touches to work with
         guard let touch = touches.first else {
             return
         }
         let touchLocation = touch.locationInNode(self)
         
-        // 2 - Set up initial location of projectile
-        let projectile = SKSpriteNode(imageNamed: "red")
-        projectile.position = player.position
-        projectile.size = CGSize(width: 10, height: 10)
-        projectile.zPosition = 1        
-        
-        // 3 - Determine offset of location to projectile
-        let offset = touchLocation - projectile.position
-        
-        // 5 - OK to add now - you've double checked position
-        addChild(projectile)
-        
-        // 6 - Get the direction of where to shoot
-        let direction = offset.normalized()
-        
-        // 7 - Make it shoot far enough to be guaranteed off screen
-        let shootAmount = direction * 1000
-        
-        // 8 - Add the shoot amount to the current position
-        let realDest = shootAmount + projectile.position
-        
-        projectile.physicsBody = SKPhysicsBody(circleOfRadius: projectile.size.width/2)
-        projectile.physicsBody?.dynamic = true
-        projectile.physicsBody?.categoryBitMask = PhysicsCategory.Projectile
-        projectile.physicsBody?.contactTestBitMask = PhysicsCategory.Enemy
-        projectile.physicsBody?.collisionBitMask = PhysicsCategory.None
-        projectile.physicsBody?.usesPreciseCollisionDetection = true
-        
-        // 9 - Create the actions
-        let actionMove = SKAction.moveTo(realDest, duration: 2.0)
-        let actionMoveDone = SKAction.removeFromParent()
-        projectile.runAction(SKAction.sequence([actionMove, actionMoveDone]))
+        // Init and add the projectile
+        let projectile = Projectile()
+        projectile.createProjectile(touchLocation, playerPosition: player.position)
+        self.addChild(projectile)
         
     }
     
     
-    func gameOver() {
-        print("test")
+    func gameOver() {        
         hud.showButtons()
     }
